Adding cache editing to mongo functions 1
